Refactor visualizer scripts to extract visualize_js into visualize_utils library
