+++ datasets/index.html
@@ -14,7 +14,7 @@
 <body>
     <div id="app"><p style="padding:2rem;font-family:sans-serif;color:#6b7280">Loading gallery…</p></div>
     <script>window.__GALLERY_DATA__ = {
-  "generated_at": "2026-04-29T12:59:54.898992Z",
+  "generated_at": "2026-06-01T06:14:10.211152Z",
   "versions": [
     {
       "label": "1.1",
