+++ python/mlcroissant/mlcroissant/scripts/visualize_utils.py
@@ -4,10 +4,13 @@ These helpers handle data previews (file listings for FileSets, sample
 rows for RecordSets) and are imported by ``visualize.py``.
 """
 
+import copy
 import fnmatch
 import json
 import pathlib
 import re
+import shutil
+import sys
 import tarfile
 from typing import Any
 import zipfile
@@ -260,3 +263,305 @@ def _get_or_generate_recordset_preview(
     for r in flat_rows:
         table_rows.append([str(r.get(c, "")) for c in cols])
     return display_cols, table_rows
+
+
+# Directory containing static assets (visualizer.js, visualizer.css)
+_STATIC_DIR = pathlib.Path(__file__).parent / "static"
+_TEMPLATES_DIR = pathlib.Path(__file__).parent / "templates"
+
+
+def _augment_distribution(
+    dist_entry: dict,
+    metadata_distribution: list,
+    folder: epath.Path | None,
+) -> dict:
+    """Add ``cr:examples`` to a distribution JSON-LD entry.
+
+    For FileObjects with a local text/CSV content URL, embeds a text preview.
+    For FileSets, embeds the resolved file listing from the parent archive.
+    Does not modify the original dict.
+    """
+    entry = copy.copy(dist_entry)
+    res_type = ""
+    t = dist_entry.get("@type", "")
+    if isinstance(t, list):
+        t = " ".join(t)
+    if "FileObject" in t or t == "sc:FileObject":
+        res_type = "FileObject"
+    elif "FileSet" in t or t == "sc:FileSet":
+        res_type = "FileSet"
+
+    if res_type == "FileObject" and folder:
+        content_url = dist_entry.get("contentUrl", "")
+        enc_fmts = dist_entry.get("encodingFormat", [])
+        if isinstance(enc_fmts, str):
+            enc_fmts = [enc_fmts]
+
+        # Case A: FileObject is a zip archive
+        if any("zip" in fmt for fmt in enc_fmts) or (
+            isinstance(content_url, str) and content_url.endswith(".zip")
+        ):
+            if content_url and not str(content_url).startswith("http"):
+                file_path = folder / content_url
+                if file_path.exists():
+                    archive_entries = _list_archive_entries(str(file_path), enc_fmts)
+                    if archive_entries:
+                        entry["cr:examples"] = {
+                            "file_list": archive_entries[:_MAX_PREVIEW_FILES],
+                            "file_count": len(archive_entries),
+                        }
+
+        # Case A (cont.): FileObject is a tar archive
+        elif any("tar" in fmt for fmt in enc_fmts) or any(
+            isinstance(content_url, str) and content_url.endswith(ext)
+            for ext in [".tar", ".tar.gz", ".tgz"]
+        ):
+            if content_url and not str(content_url).startswith("http"):
+                file_path = folder / content_url
+                if file_path.exists():
+                    archive_entries = _list_archive_entries(str(file_path), enc_fmts)
+                    if archive_entries:
+                        entry["cr:examples"] = {
+                            "file_list": archive_entries[:_MAX_PREVIEW_FILES],
+                            "file_count": len(archive_entries),
+                        }
+
+        # Case B: FileObject is contained within a zip archive
+        contained_in = dist_entry.get("containedIn")
+        if contained_in:
+            if isinstance(contained_in, list):
+                contained_in = contained_in[0] if contained_in else None
+
+            if isinstance(contained_in, str):
+                # Find parent resource
+                parent = None
+                for res in metadata_distribution:
+                    if res.name == contained_in or res.uuid == contained_in:
+                        parent = res
+                        break
+
+                if parent and hasattr(parent, "content_url") and parent.content_url:
+                    parent_url = str(parent.content_url)
+                    if not parent_url.startswith("http"):
+                        parent_path = folder / parent_url
+                        if parent_path.exists():
+                            parent_enc = getattr(parent, "encoding_formats", []) or []
+                            if isinstance(parent_enc, str):
+                                parent_enc = [parent_enc]
+
+                            if any(
+                                "zip" in fmt for fmt in parent_enc
+                            ) or parent_url.endswith(".zip"):
+                                try:
+                                    with zipfile.ZipFile(str(parent_path)) as z:
+                                        if content_url in z.namelist():
+                                            with z.open(content_url) as f:
+                                                # Read first 5 lines
+                                                lines = []
+                                                for _ in range(5):
+                                                    line = f.readline()
+                                                    if not line:
+                                                        break
+                                                    lines.append(
+                                                        line.decode(
+                                                            "utf-8", errors="replace"
+                                                        )
+                                                    )
+                                                text_preview = "".join(lines)
+                                                entry["cr:examples"] = {
+                                                    "text_preview": text_preview
+                                                }
+                                except Exception as e:
+                                    logging.warning(
+                                        f"Failed to read from zip {parent_path}: {e}"
+                                    )
+                            elif (
+                                any("tar" in str(fmt).lower() for fmt in parent_enc)
+                                or any(fmt == "application/x-tar" for fmt in parent_enc)
+                                or any(
+                                    parent_url.endswith(ext)
+                                    for ext in [".tar", ".tar.gz", ".tgz"]
+                                )
+                            ):
+                                try:
+                                    with tarfile.open(str(parent_path)) as t:
+                                        try:
+                                            member = t.getmember(content_url)
+                                            f_obj = t.extractfile(member)
+                                            if f_obj is not None:
+                                                # Read first 5 lines
+                                                lines = []
+                                                for _ in range(5):
+                                                    line = f_obj.readline()
+                                                    if not line:
+                                                        break
+                                                    lines.append(
+                                                        line.decode(
+                                                            "utf-8", errors="replace"
+                                                        )
+                                                    )
+                                                text_preview = "".join(lines)
+                                                entry["cr:examples"] = {
+                                                    "text_preview": text_preview
+                                                }
+                                        except KeyError:
+                                            logging.warning(
+                                                f"File {content_url} not found in tar"
+                                                f" {parent_path}"
+                                            )
+                                except Exception as e:
+                                    logging.warning(
+                                        f"Failed to read from tar {parent_path}: {e}"
+                                    )
+
+        # Original Case: Direct local file
+        elif content_url and not str(content_url).startswith("http"):
+            file_path = folder / content_url
+            if file_path.exists():
+                if (
+                    enc_fmts
+                    and any(
+                        fmt in ["text/csv", "text/plain", "application/json"]
+                        for fmt in enc_fmts
+                    )
+                ) or (isinstance(content_url, str) and content_url.endswith(".json")):
+                    try:
+                        with file_path.open("r") as f:
+                            text_preview = "".join([f.readline() for _ in range(5)])
+                        entry["cr:examples"] = {"text_preview": text_preview}
+                    except Exception as e:
+                        logging.warning(
+                            f"Could not read preview for {dist_entry.get('name')}: {e}"
+                        )
+
+    if res_type == "FileSet" and folder:
+        # We need an mlcroissant resource object to resolve the fileset.
+        # Find the matching resource from metadata.distribution.
+        name = dist_entry.get("name") or dist_entry.get("@id") or ""
+        matching_res = None
+        for res in metadata_distribution:
+            if res.name == name or res.uuid == name:
+                matching_res = res
+                break
+        if matching_res is not None:
+            file_list, file_count = _resolve_fileset_files(
+                matching_res, metadata_distribution, folder
+            )
+            includes = list(getattr(matching_res, "includes", []) or [])
+            entry["cr:examples"] = {
+                "file_list": file_list,
+                "file_count": file_count,
+                "includes": includes,
+            }
+
+    return entry
+
+
+def _augment_record_set(
+    rs_entry: dict,
+    dataset: mlc.Dataset,
+    folder: epath.Path | None,
+) -> dict:
+    """Add ``cr:examples`` to a record set JSON-LD entry with preview data."""
+    entry = copy.copy(rs_entry)
+    rs_name = rs_entry.get("name") or rs_entry.get("@id") or ""
+    if not rs_name:
+        return entry
+    preview_cols, preview_rows = _get_or_generate_recordset_preview(
+        dataset, rs_name, folder
+    )
+    if preview_cols and preview_rows:
+        entry["cr:examples"] = {
+            "columns": preview_cols,
+            "rows": preview_rows,
+        }
+    return entry
+
+
+def visualize_js(
+    jsonld: str,
+    output: epath.Path,
+    static_path: str = ".",
+    gallery_url: str | None = None,
+) -> None:
+    """Generate a JS-driven visualization HTML page for a Croissant dataset.
+
+    Args:
+        jsonld: Path to the Croissant JSON-LD file.
+        output: Output HTML file path.
+        static_path: Path or URL to the js/css directory.
+        gallery_url: Optional relative URL back to the gallery index (e.g.
+            ``../../index.html``). When set, the visualizer renders a
+            "\u2190 All Datasets" back-link in the sidebar. Only passed by
+            ``visualize_all.py``; defaults to None so that single-dataset
+            local renders are unaffected.
+    """
+    logging.info(f"Loading dataset from {jsonld}")
+    try:
+        dataset = mlc.Dataset(jsonld)
+    except Exception as e:
+        logging.error(f"Failed to load dataset: {e}")
+        sys.exit(1)
+
+    metadata = dataset.metadata
+    folder: epath.Path | None = metadata.ctx.folder
+
+    # Load raw JSON-LD
+    jsonld_path = epath.Path(jsonld)
+    try:
+        raw_jsonld: dict[str, Any] = json.loads(jsonld_path.read_text())
+    except Exception:
+        raw_jsonld = {}
+
+    # Deep-copy so we can augment without mutating the original
+    augmented = copy.deepcopy(raw_jsonld)
+
+    # Augment distributions with cr:examples
+    distributions = augmented.get("distribution", [])
+    if not isinstance(distributions, list):
+        distributions = [distributions]
+    augmented_dists = [
+        _augment_distribution(d, metadata.distribution, folder) for d in distributions
+    ]
+    augmented["distribution"] = augmented_dists
+
+    # Augment record sets with cr:examples
+    record_sets = augmented.get("recordSet", [])
+    if not isinstance(record_sets, list):
+        record_sets = [record_sets]
+    augmented_rs = [_augment_record_set(rs, dataset, folder) for rs in record_sets]
+    augmented["recordSet"] = augmented_rs
+
+    # Inject the augmented JSON-LD inline into the HTML template.
+    # A simple placeholder replacement keeps this Jinja2-free.
+    # The inline <script> sets window.__CROISSANT_DATA__ synchronously,
+    # so the page works from file:// with no web server required.
+    output_dir = pathlib.Path(str(output)).parent
+    data_script = (
+        "<script>window.__CROISSANT_DATA__ = "
+        + json.dumps(augmented, indent=2, ensure_ascii=False)
+        + ";</script>"
+    )
+    # Optional back-link to gallery index (only injected by visualize_all.py).
+    gallery_script = (
+        f"<script>window.__GALLERY_URL__={json.dumps(gallery_url)};</script>"
+        if gallery_url is not None
+        else ""
+    )
+    template_html = (_TEMPLATES_DIR / "visualizer.html").read_text(encoding="utf-8")
+    html = template_html.replace("<!-- DATASET_DATA -->", data_script)
+    html = html.replace("<!-- GALLERY_URL -->", gallery_script)
+    html = html.replace("STATIC_PATH", static_path)
+    pathlib.Path(str(output)).write_text(html, encoding="utf-8")
+    print(f"Wrote visualization to {output}")
+
+    # Copy static assets next to the output file if using default path
+    if static_path == ".":
+        for asset in ("visualizer.js", "visualizer.css"):
+            src = _STATIC_DIR / asset
+            dst = output_dir / asset
+            if src.exists():
+                shutil.copy(src, dst)
+                logging.info(f"Copied {src} -> {dst}")
+            else:
+                logging.warning(f"Static asset not found: {src}")
